game presenter разбит на разные классы,чтобы разделять ответственность за разные модели
