+++ main.py
@@ -1,5 +1,5 @@
 import curses
-from Presenters.GamePresenter import GamePresenter
+from Presenters.RootPresenter import RootPresenter
 import locale
 
 
@@ -7,7 +7,7 @@ import locale
 def main(stdscr: curses.window):
     stdscr.encoding = 'utf-8'
     locale.setlocale(locale.LC_ALL, 'zh_CN.UTF-8')
-    game_presenter = GamePresenter(stdscr)
-    game_presenter.run()
+    root_presenter = RootPresenter(stdscr)
+    root_presenter.run()
 
 curses.wrapper(main)
