+++ records.json
@@ -974,6 +974,253 @@
             "wpm": 16.25,
             "time_elapsed": 36.93,
             "completed": true
+        },
+        {
+            "timestamp": "2025-04-26 14:29:21",
+            "language": "Chinese",
+            "difficulty": "simple",
+            "level": "Level 0",
+            "chars_typed": 15,
+            "total_chars": 30,
+            "correct_keystrokes": 14,
+            "accuracy": 93.33,
+            "wpm": 3.0,
+            "time_elapsed": 60.0,
+            "completed": false
+        },
+        {
+            "timestamp": "2025-04-26 14:33:29",
+            "language": "English",
+            "difficulty": "simple",
+            "level": "Level 0",
+            "chars_typed": 93,
+            "total_chars": 93,
+            "correct_keystrokes": 16,
+            "accuracy": 17.2,
+            "wpm": 169.93,
+            "time_elapsed": 6.57,
+            "completed": true
+        },
+        {
+            "timestamp": "2025-04-26 14:34:35",
+            "language": "Chinese",
+            "difficulty": "simple",
+            "level": "Level 0",
+            "chars_typed": 13,
+            "total_chars": 30,
+            "correct_keystrokes": 12,
+            "accuracy": 92.31,
+            "wpm": 2.6,
+            "time_elapsed": 60.01,
+            "completed": false
+        },
+        {
+            "timestamp": "2025-04-26 14:41:05",
+            "language": "Chinese",
+            "difficulty": "simple",
+            "level": "Level 0",
+            "chars_typed": 18,
+            "total_chars": 30,
+            "correct_keystrokes": 13,
+            "accuracy": 72.22,
+            "wpm": 3.6,
+            "time_elapsed": 60.0,
+            "completed": false
+        },
+        {
+            "timestamp": "2025-04-26 14:43:56",
+            "language": "English",
+            "difficulty": "middle",
+            "level": "Level 0",
+            "chars_typed": 134,
+            "total_chars": 134,
+            "correct_keystrokes": 118,
+            "accuracy": 88.06,
+            "wpm": 30.8,
+            "time_elapsed": 52.2,
+            "completed": true
+        },
+        {
+            "timestamp": "2025-04-26 14:46:38",
+            "language": "English",
+            "difficulty": "middle",
+            "level": "Level 1",
+            "chars_typed": 129,
+            "total_chars": 129,
+            "correct_keystrokes": 29,
+            "accuracy": 22.48,
+            "wpm": 54.18,
+            "time_elapsed": 28.57,
+            "completed": true
+        },
+        {
+            "timestamp": "2025-04-26 14:49:45",
+            "language": "Chinese",
+            "difficulty": "simple",
+            "level": "Level 0",
+            "chars_typed": 15,
+            "total_chars": 30,
+            "correct_keystrokes": 15,
+            "accuracy": 100.0,
+            "wpm": 3.0,
+            "time_elapsed": 60.01,
+            "completed": false
+        },
+        {
+            "timestamp": "2025-04-26 14:50:15",
+            "language": "Russian",
+            "difficulty": "middle",
+            "level": "Level 1",
+            "chars_typed": 164,
+            "total_chars": 164,
+            "correct_keystrokes": 10,
+            "accuracy": 6.1,
+            "wpm": 118.8,
+            "time_elapsed": 16.57,
+            "completed": true
+        },
+        {
+            "timestamp": "2025-04-26 15:14:01",
+            "language": "Chinese",
+            "difficulty": "simple",
+            "level": "Level 0",
+            "chars_typed": 15,
+            "total_chars": 30,
+            "correct_keystrokes": 12,
+            "accuracy": 80.0,
+            "wpm": 3.0,
+            "time_elapsed": 60.0,
+            "completed": false
+        },
+        {
+            "timestamp": "2025-04-26 15:49:23",
+            "language": "English",
+            "difficulty": "simple",
+            "level": "Level 0",
+            "chars_typed": 92,
+            "total_chars": 93,
+            "correct_keystrokes": 60,
+            "accuracy": 65.22,
+            "wpm": 18.4,
+            "time_elapsed": 60.0,
+            "completed": false
+        },
+        {
+            "timestamp": "2025-04-26 15:52:18",
+            "language": "English",
+            "difficulty": "middle",
+            "level": "Level 2",
+            "chars_typed": 155,
+            "total_chars": 155,
+            "correct_keystrokes": 36,
+            "accuracy": 23.23,
+            "wpm": 80.81,
+            "time_elapsed": 23.02,
+            "completed": true
+        },
+        {
+            "timestamp": "2025-04-26 15:53:57",
+            "language": "Russian",
+            "difficulty": "simple",
+            "level": "Level 0",
+            "chars_typed": 88,
+            "total_chars": 88,
+            "correct_keystrokes": 27,
+            "accuracy": 30.68,
+            "wpm": 28.53,
+            "time_elapsed": 37.02,
+            "completed": true
+        },
+        {
+            "timestamp": "2025-04-26 15:57:18",
+            "language": "Russian",
+            "difficulty": "middle",
+            "level": "Level 2",
+            "chars_typed": 125,
+            "total_chars": 153,
+            "correct_keystrokes": 118,
+            "accuracy": 94.4,
+            "wpm": 25.0,
+            "time_elapsed": 60.0,
+            "completed": false
+        },
+        {
+            "timestamp": "2025-04-26 15:58:28",
+            "language": "Russian",
+            "difficulty": "middle",
+            "level": "Level 3",
+            "chars_typed": 129,
+            "total_chars": 136,
+            "correct_keystrokes": 98,
+            "accuracy": 75.97,
+            "wpm": 25.8,
+            "time_elapsed": 60.0,
+            "completed": false
+        },
+        {
+            "timestamp": "2025-04-26 15:59:45",
+            "language": "Chinese",
+            "difficulty": "simple",
+            "level": "Level 0",
+            "chars_typed": 13,
+            "total_chars": 30,
+            "correct_keystrokes": 10,
+            "accuracy": 76.92,
+            "wpm": 2.6,
+            "time_elapsed": 60.0,
+            "completed": false
+        },
+        {
+            "timestamp": "2025-05-01 03:03:05",
+            "language": "English",
+            "difficulty": "middle",
+            "level": "Level 1",
+            "chars_typed": 77,
+            "total_chars": 129,
+            "correct_keystrokes": 68,
+            "accuracy": 88.31,
+            "wpm": 15.4,
+            "time_elapsed": 60.0,
+            "completed": false
+        },
+        {
+            "timestamp": "2025-05-01 03:04:40",
+            "language": "Russian",
+            "difficulty": "hard",
+            "level": "Level 4",
+            "chars_typed": 87,
+            "total_chars": 180,
+            "correct_keystrokes": 79,
+            "accuracy": 90.8,
+            "wpm": 17.4,
+            "time_elapsed": 60.01,
+            "completed": false
+        },
+        {
+            "timestamp": "2025-05-01 03:05:46",
+            "language": "Chinese",
+            "difficulty": "middle",
+            "level": "Level 1",
+            "chars_typed": 19,
+            "total_chars": 19,
+            "correct_keystrokes": 2,
+            "accuracy": 10.53,
+            "wpm": 7.88,
+            "time_elapsed": 28.95,
+            "completed": true
+        },
+        {
+            "timestamp": "2025-05-01 03:06:29",
+            "language": "English",
+            "difficulty": "simple",
+            "level": "Level 1",
+            "chars_typed": 74,
+            "total_chars": 74,
+            "correct_keystrokes": 25,
+            "accuracy": 33.78,
+            "wpm": 25.13,
+            "time_elapsed": 35.33,
+            "completed": true
         }
     ]
 }
